// src/app/shared/models/interfaces/boats.interface.ts
export interface Boats {
  vesselId: number;
  externalId: number;
  name: string;
  referenceNumber: string;
  type: string;
  yearOfBuilt: number;
  capacity: number;
  tonnage: number;
  length: number;
  width: number;
  draft: number;
  speed: number;
  description: string;
  status: string;
  createdDate: string;
  lastModifiedDate: string;
}